Abstracted Logic Code away from Custom Event Definition
I moved the logic code away from the custom event definition I made and abstracted it all in a method. This makes implementation easy as there is no work to do outside of instantiating the class on your window.

Originally I wanted to leave the logic obviously open for devs, but that made project implementation harder than needed.